Fix: Skips conversion to WAV if already detected as WAV
